Make some adjustments
