--- __tests__/main.test.ts
@@ -1,10 +1,0 @@
-// Import { getToken } from '../src/get-token'
-
-// Parameters for getToken include the private key from the Github ok-to-test App,
-// the app id that can be found in the Github App UI, the owner of the repo, and the repo
-// name where the app is installed
-// test('generate token', async () => {
-//   const input = await getToken(
-//     ``, "", { owner: "", repo: "" })
-//   console.log(input)
-// })
