--- jest.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  clearMocks: true,
-  moduleFileExtensions: ["js", "ts"],
-  preset: "ts-jest",
-  testEnvironment: "node",
-  testMatch: ["**/*.test.ts"],
-  verbose: true,
-};
