+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "github-app-token",
-  "version": "1.1.4",
+  "version": "1.2.0",
   "license": "MIT",
   "files": [
     "action.yml",
@@ -12,34 +12,30 @@
     "check-prettier": "yarn run prettier --check",
     "eslint": "eslint --ignore-path .gitignore --max-warnings 0 \"./**/*.{js,ts}\"",
     "format-prettier": "yarn run prettier --write",
-    "prettier": "prettier --ignore-path .gitignore \"./**/*.{js,json,md,ts,yml}\"",
-    "test": "jest"
+    "prettier": "prettier --ignore-path .gitignore \"./**/*.{js,json,md,ts,yml}\""
   },
   "devDependencies": {
     "@actions/core": "^1.2.6",
     "@actions/github": "^4.0.0",
-    "@octokit/auth-app": "^2.10.4",
+    "@octokit/auth-app": "^2.10.5",
     "@types/is-base64": "^1.1.0",
     "@types/jest": "^26.0.18",
-    "@types/node": "^14.14.6",
-    "@typescript-eslint/eslint-plugin": "^4.6.0",
-    "@typescript-eslint/parser": "^4.6.0",
-    "@vercel/ncc": "^0.24.1",
-    "eslint": "^7.12.1",
-    "eslint-config-prettier": "^6.15.0",
+    "@types/node": "^14.14.11",
+    "@typescript-eslint/eslint-plugin": "^4.9.1",
+    "@typescript-eslint/parser": "^4.9.1",
+    "@vercel/ncc": "^0.25.1",
+    "eslint": "^7.15.0",
+    "eslint-config-prettier": "^7.0.0",
     "eslint-config-xo": "^0.33.1",
-    "eslint-config-xo-typescript": "^0.35.0",
+    "eslint-config-xo-typescript": "^0.36.0",
     "eslint-import-resolver-typescript": "^2.3.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-sort-destructure-keys": "^1.3.5",
     "eslint-plugin-typescript-sort-keys": "^1.5.0",
     "eslint-plugin-unicorn": "^23.0.0",
     "is-base64": "^1.1.0",
-    "jest": "^26.6.3",
-    "prettier": "^2.1.2",
+    "prettier": "^2.2.1",
     "promise-retry": "^2.0.1",
-    "semver": "^7.3.4",
-    "ts-jest": "^26.4.4",
-    "typescript": "^4.0.5"
+    "typescript": "^4.1.2"
   }
 }
