+++ src/index.ts
@@ -1,39 +1,31 @@
-import {
-  error as logError,
-  getInput,
-  info,
-  setFailed,
-  setOutput,
-  setSecret,
-} from "@actions/core";
-import { context, getOctokit } from "@actions/github";
+import { getInput, info, setFailed, setOutput, setSecret } from "@actions/core";
 
 import isBase64 from "is-base64";
 
-import { getToken } from "./get-token";
+import { fetchInstallationToken } from "./fetch-installation-token";
 
 const run = async () => {
   try {
-    // Parse inputs
-    const id = getInput("app_id", { required: true });
+    const appId = getInput("app_id", { required: true });
     const privateKeyInput = getInput("private_key", { required: true });
     const privateKey = isBase64(privateKeyInput)
       ? Buffer.from(privateKeyInput, "base64").toString("utf8")
       : privateKeyInput;
 
-    // Run our actual logic
-    const installationToken = await getToken(privateKey, id, context.repo);
+    const installationToken = await fetchInstallationToken({
+      appId,
+      privateKey,
+    });
 
-    // Set our inputs as needed
     setSecret(installationToken);
     setOutput("token", installationToken);
     info("Token generated successfully!");
   } catch (error: unknown) {
-    if (typeof error !== "string" && !(error instanceof Error)) {
-      throw new TypeError(`Caught error of unexpected type: ${typeof error}`);
+    if (typeof error === "string" || error instanceof Error) {
+      setFailed(error);
+    } else {
+      setFailed(`Caught error of unexpected type: ${typeof error}`);
     }
-
-    setFailed(error);
   }
 };
